немного доработки
