+++ index.d.ts
@@ -1,3 +1,9 @@
 // index.d.ts
 declare module "*.jpg";
 declare module "*.jpeg";
+
+export {};
+
+declare global {
+    interface Window { application: any; }
+}
