+++ tsconfig.json
@@ -7,6 +7,7 @@
         "target": "es5",
         "allowJs": true,
         "sourceMap": true,
-        "moduleResolution": "node"
+        "moduleResolution": "node",
+        "strictNullChecks": false,
     }
 }
